refactor: MemberService 분리 (#175)
