+++ core/network/src/main/java/com/easyhz/noffice/core/network/di/service/AuthServiceModule.kt
@@ -2,6 +2,7 @@ package com.easyhz.noffice.core.network.di.service
 
 import com.easyhz.noffice.core.network.api.auth.AuthService
 import com.easyhz.noffice.core.network.api.auth.TokenService
+import com.easyhz.noffice.core.network.api.member.MemberService
 import com.easyhz.noffice.core.network.di.NofficeRetrofit
 import com.easyhz.noffice.core.network.di.TokenRetrofit
 import dagger.Module
@@ -20,4 +21,8 @@ object AuthServiceModule {
     @Provides
     fun provideTokeService(@TokenRetrofit retrofit: Retrofit): TokenService =
         retrofit.create(TokenService::class.java)
+
+    @Provides
+    fun provideMemberService(@NofficeRetrofit retrofit: Retrofit): MemberService =
+        retrofit.create(MemberService::class.java)
 }
